Add CONTRIBUTING.md and update CODE-OF-CONDUCT.md
This patch adds a CONTRIBUTING file to the repository, and
updates the existing code of conduct to a new version which
extends the scope of the document and includes new wording.

As a part of this patch, the README file is updated to include
a reference to the CLA process.

Additionally copyright headers are also added to auto-generated
boilerplate files.