Automate generation of antrea manifest template
antrea-manifest/antrea.yml is a template file which is based on contents
from Antrea repository.

This template file should be in sync with the contents in the antrea
repo, and this PR automates the synchronization of the file with the
Antrea repo.

To generate the manifest, use hack/generate-manifest.sh
Alternately use make antrea-manifest target to build the manifest and
put it in place into antrea-manifest dir.

Signed-off-by: Kobi Samoray <ksamoray@vmware.com>