Generate deploy/*/role.yaml
Deployment requires appropriate role assignment for the deployer.
In addition to the permissions which are required to deploy, the
deployer requires all the permissions required by Antrea itself, as it
assigns the permissions to Antrea Agent and Controller.

As the required permissions by Antrea change occasionally, it's better
to automate the creation of the role.yaml files which contain the list
of required roles.

Signed-off-by: Kobi Samoray <ksamoray@vmware.com>