Certification bundle generation makefile changes
Bundle for RedHat certification requires use of the image hash instead
of image version.
This requires a build of the Antrea Operator for the given version -
which requires GitHub tagging of the repo.
But then the tagged repo will not include the set of yamls for the given
release.
So as a solution this change splits the generation process into two
phases:
- Run the following before tagging the build and include their outputs
  in the release:
 make bundle
 make ocpbundle

These will generate yamls with the operator image tagged by the version.

- Run the following after the repo is tagged and the operator image
  exists for the given release:

  make ocpcertification

Which will generate the yamls with the operator image identified by the
hash.

Signed-off-by: Kobi Samoray <ksamoray@vmware.com>