Generate human-readable antrea-config
By default, PyYAML chooses the style of a collection depending on whether it has nested collections.
If collections are always serialized in the block style, set the parameter default_flow_style of
dump() to False. And set allow_unicode to True as there are some special character like `µs` in
antrea-config. Also, generate antrea resources according to latest commit on main branch of Antrea.