Install Golang 1.19 for builds
Starting with Go 1.21, the runtime system needs to have a version of glibc "compatible" with the version available at build time. The ubi8 image comes with a very old version of glibc compared to the golang builder. so, the Antrea Operator binaries will fail to run and hit GLIBC issue.

To fix this issue, we need to install Golang manually in ubi8 image when making docker build. Moreover, we use Golang 1.19 for Github workflow build action. To align this workflow, we install Golang 1.19 for Github workflow build tag action and Docker image build.

Also upgrade GitHub actions to latest versions due to node deprecation
warning.

Signed-off-by: Kobi Samoray <ksamoray@vmware.com>