Replace version/version.go with internal/version/version.go
On case-insensitive filesystems (which includes default macOS
installations), it is not possible to have a "version" directory and a
"VERSION" file in the same parent directory, and the repo cannot be
cloned succesfully. To resolve this, we replace version/version.go with
internal/version/version.go.

Additionally, we stop using "0.0.1" as the version string, as it is not
consistent with the contents of the VERSION file, nor with the Docker
tag. Instead, we use the VERSION file as the one source of truth. We
also add a "--version" command-line flag to the operator binary, which
displays the version and exits.

Signed-off-by: Antonin Bas <abas@vmware.com>