Uniform tool installation in Makefile
Use a uniform installation process for golangci-lint, controller-gen and
kustomize.

The process ensures that the correct version of the tool is used for
every "build", regardless of what other version may already be installed
on the system.

For kustomize, we ensure that the Makefile is the single source of
truth, and we remove hack/verify-kustomize.sh.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>